+++ inlang/source-code/ide-extension/assets/settings-component.mjs
@@ -26971,7 +26971,7 @@ var b = class extends HTMLElement {
 };
 b.elementStyles = [], b.shadowRootOptions = { mode: "open" }, b[d("elementProperties")] = /* @__PURE__ */ new Map(), b[d("finalized")] = /* @__PURE__ */ new Map(), p?.({ ReactiveElement: b }), (a.reactiveElementVersions ??= []).push("2.0.4");
 
-// ../../../node_modules/.pnpm/lit-html@3.1.3/node_modules/lit-html/lit-html.js
+// ../../../node_modules/.pnpm/lit-html@3.1.4/node_modules/lit-html/lit-html.js
 var t2 = globalThis;
 var i3 = t2.trustedTypes;
 var s2 = i3 ? i3.createPolicy("lit-html", { createHTML: (t7) => t7 }) : void 0;
@@ -27229,7 +27229,7 @@ var L = class {
 };
 var z = { P: e3, A: h2, C: o3, M: 1, L: P, R: S2, D: u2, V: N, I: M, H: R, N: H, U: I, B: k, F: L };
 var Z = t2.litHtmlPolyfillSupport;
-Z?.(V, M), (t2.litHtmlVersions ??= []).push("3.1.3");
+Z?.(V, M), (t2.litHtmlVersions ??= []).push("3.1.4");
 var j = (t7, i5, s5) => {
   const e11 = s5?.renderBefore ?? i5;
   let h3 = e11._$litPart$;
@@ -28131,7 +28131,7 @@ var InlangModule = import_typebox9.Type.Object({
 // ../sdk/dist/resolve-modules/message-lint-rules/resolveMessageLintRules.js
 var import_value = __toESM(require_value2(), 1);
 
-// ../../../node_modules/.pnpm/dedent@1.5.1/node_modules/dedent/dist/dedent.mjs
+// ../../../node_modules/.pnpm/dedent@1.5.1_babel-plugin-macros@2.8.0/node_modules/dedent/dist/dedent.mjs
 function ownKeys(object, enumerableOnly) {
   var keys = Object.keys(object);
   if (Object.getOwnPropertySymbols) {
@@ -29567,6 +29567,7 @@ var registry = [
         "/": "./inlang/source-code/paraglide/paraglide-js/docs/why-paraglide.md",
         "/getting-started": "./inlang/source-code/paraglide/paraglide-js/README.md",
         "/usage": "./inlang/source-code/paraglide/paraglide-js/docs/usage.md",
+        "/scaling": "./inlang/source-code/paraglide/paraglide-js/docs/scaling.md",
         "/architecture": "./inlang/source-code/paraglide/paraglide-js/docs/architecture.md",
         "/changelog": "./inlang/source-code/paraglide/paraglide-js/CHANGELOG.md"
       },
@@ -29782,8 +29783,7 @@ var registry = [
         "/": "./inlang/source-code/paraglide/paraglide-sveltekit/docs/why-paraglide.md",
         "/getting-started": "./inlang/source-code/paraglide/paraglide-sveltekit/README.md",
         "/localised-routing": "./inlang/source-code/paraglide/paraglide-sveltekit/docs/localised-routing.md",
-        "/advanced-usage": "./inlang/source-code/paraglide/paraglide-sveltekit/docs/advanced-usage.md",
-        "/serverside-usage": "./inlang/source-code/paraglide/paraglide-sveltekit/docs/serverside-usage.md"
+        "/advanced-usage": "./inlang/source-code/paraglide/paraglide-sveltekit/docs/advanced-usage.md"
       },
       Appendix: {
         "/manual-setup": "./inlang/source-code/paraglide/paraglide-sveltekit/docs/manual-setup.md",
@@ -32300,7 +32300,7 @@ __decorateClass2([
   n4()
 ], ShoelaceElement.prototype, "lang", 2);
 
-// ../../../node_modules/.pnpm/lit-html@3.1.3/node_modules/lit-html/directive-helpers.js
+// ../../../node_modules/.pnpm/lit-html@3.1.4/node_modules/lit-html/directive-helpers.js
 var { I: t4 } = z;
 var e6 = (o9, t7) => void 0 === t7 ? void 0 !== o9?._$litType$ : o9?._$litType$ === t7;
 var f3 = (o9) => void 0 === o9.strings;
@@ -32457,7 +32457,7 @@ __decorateClass2([
   watch(["name", "src", "library"])
 ], SlIcon.prototype, "setIcon", 1);
 
-// ../../../node_modules/.pnpm/lit-html@3.1.3/node_modules/lit-html/directive.js
+// ../../../node_modules/.pnpm/lit-html@3.1.4/node_modules/lit-html/directive.js
 var t5 = { ATTRIBUTE: 1, CHILD: 2, PROPERTY: 3, BOOLEAN_ATTRIBUTE: 4, EVENT: 5, ELEMENT: 6 };
 var e7 = (t7) => (...e11) => ({ _$litDirective$: t7, values: e11 });
 var i4 = class {
@@ -32477,7 +32477,7 @@ var i4 = class {
   }
 };
 
-// ../../../node_modules/.pnpm/lit-html@3.1.3/node_modules/lit-html/directives/class-map.js
+// ../../../node_modules/.pnpm/lit-html@3.1.4/node_modules/lit-html/directives/class-map.js
 var e8 = e7(class extends i4 {
   constructor(t7) {
     if (super(t7), t7.type !== t5.ATTRIBUTE || "class" !== t7.name || t7.strings?.length > 2)
@@ -32504,7 +32504,7 @@ var e8 = e7(class extends i4 {
   }
 });
 
-// ../../../node_modules/.pnpm/lit-html@3.1.3/node_modules/lit-html/static.js
+// ../../../node_modules/.pnpm/lit-html@3.1.4/node_modules/lit-html/static.js
 var e9 = Symbol.for("");
 var o5 = (t7) => {
   if (t7?.r === e9)
@@ -32536,7 +32536,7 @@ var l3 = (t7) => (r8, ...e11) => {
 var n5 = l3(x);
 var u4 = l3(b2);
 
-// ../../../node_modules/.pnpm/lit-html@3.1.3/node_modules/lit-html/directives/if-defined.js
+// ../../../node_modules/.pnpm/lit-html@3.1.4/node_modules/lit-html/directives/if-defined.js
 var o6 = (o9) => o9 ?? T;
 
 // ../../../node_modules/.pnpm/@shoelace-style+shoelace@2.14.0_@types+react@18.3.1/node_modules/@shoelace-style/shoelace/dist/chunks/chunk.7XLSSP47.js
@@ -35561,7 +35561,7 @@ var HasSlotController = class {
   }
 };
 
-// ../../../node_modules/.pnpm/lit-html@3.1.3/node_modules/lit-html/directives/unsafe-html.js
+// ../../../node_modules/.pnpm/lit-html@3.1.4/node_modules/lit-html/directives/unsafe-html.js
 var e10 = class extends i4 {
   constructor(i5) {
     if (super(i5), this.it = T, i5.type !== t5.CHILD)
@@ -36796,7 +36796,7 @@ var input_styles_default = i`
   }
 `;
 
-// ../../../node_modules/.pnpm/lit-html@3.1.3/node_modules/lit-html/directives/live.js
+// ../../../node_modules/.pnpm/lit-html@3.1.4/node_modules/lit-html/directives/live.js
 var l4 = e7(class extends i4 {
   constructor(r8) {
     if (super(r8), r8.type !== t5.PROPERTY && r8.type !== t5.ATTRIBUTE && r8.type !== t5.BOOLEAN_ATTRIBUTE)