+++ inlang/source-code/editor/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inlang/editor
 
+## 0.1.6
+
+### Patch Changes
+
+- @inlang/settings-component@1.0.29
+
+## 0.1.5
+
+### Patch Changes
+
+- @inlang/settings-component@1.0.28
+
+## 0.1.4
+
+### Patch Changes
+
+- @inlang/settings-component@1.0.27
+
 ## 0.1.3
 
 ### Patch Changes