+++ inlang/source-code/end-to-end-tests/paraglide-next/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @inlang/paraglide-next-e2e
 
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+  - @inlang/paraglide-next@0.4.4
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [94f365b]
+  - @inlang/paraglide-next@0.4.4
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+  - @inlang/paraglide-next@0.4.3
+
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+  - @inlang/paraglide-next@0.4.2
+
 ## 0.0.19
 
 ### Patch Changes