+++ inlang/source-code/github-lint-action/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/github-lint-action
 
+## 0.3.16
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+
+## 0.3.15
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+
+## 0.3.14
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
 ## 0.3.13
 
 ### Patch Changes