+++ inlang/source-code/paraglide/paraglide-astro/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inlang/paraglide-astro
 
+## 0.1.31
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.71
+
+## 0.1.30
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.70
+
+## 0.1.29
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.69
+
 ## 0.1.28
 
 ### Patch Changes