+++ inlang/source-code/paraglide/paraglide-sveltekit/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @inlang/paraglide-sveltekit
 
+## 0.10.6
+
+### Patch Changes
+
+- bcd894f: `init` command now generates the `src/lib` directory if it does not exist. Previously it would just crash
+  - @inlang/paraglide-js@1.10.1
+  - @inlang/paraglide-vite@1.2.71
+
+## 0.10.5
+
+### Patch Changes
+
+- Updated dependencies [c5d145d]
+  - @inlang/paraglide-js@1.10.1
+  - @inlang/paraglide-vite@1.2.70
+
+## 0.10.4
+
+### Patch Changes
+
+- Updated dependencies [33662e6]
+  - @inlang/paraglide-js@1.10.0
+  - @inlang/paraglide-vite@1.2.69
+
 ## 0.10.3
 
 ### Patch Changes