+++ inlang/source-code/plugins/i18next/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/plugin-i18next
 
+## 4.14.12
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+
+## 4.14.11
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+
+## 4.14.10
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
 ## 4.14.9
 
 ### Patch Changes