+++ inlang/source-code/settings-component/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/settings-component
 
+## 1.0.29
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+
+## 1.0.28
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+
+## 1.0.27
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
 ## 1.0.26
 
 ### Patch Changes