+++ inlang/source-code/telemetry/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/telemetry
 
+## 0.3.49
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+
+## 0.3.48
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+
+## 0.3.47
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
 ## 0.3.46
 
 ### Patch Changes